global: Fix some warnings

+++ src/common/texture.cpp
@@ -12,8 +12,8 @@ void FlipRGBA8Texture(std::vector<u8>& tex, u64 width, u64 height) {
     ASSERT(tex.size() == width * height * 4);
     const u64 line_size = width * 4;
     for (u64 line = 0; line < height / 2; line++) {
-        const u32 offset_1 = line * line_size;
-        const u32 offset_2 = (height - line - 1) * line_size;
+        const u64 offset_1 = line * line_size;
+        const u64 offset_2 = (height - line - 1) * line_size;
         // Swap lines
         std::swap_ranges(tex.begin() + offset_1, tex.begin() + offset_1 + line_size,
                          tex.begin() + offset_2);
